requirements.txt
fire~=0.7.0
uvicorn[standard]~=0.22.0
python-dotenv~=1.0.1
dhooks~=1.1.4
loguru~=0.7.2
fastapi~=0.99.0
orjson~=3.10.12
aiohttp~=3.10.11
httpx~=0.23.3
pandas~=2.2.3
ebest